Remove unused code. Fix drain method definition.
Code that was left behind as went through and reworked functions was removed:
`immediate-close`.

A drain will optionally get an entry, depends on if it's told to close or
flush, which don't require entries. Making it optional gives it the default of
`nil` which is what drains should be looking for anyway for indication as to
an entry's presence.

FossilOrigin-Name: a55336cede135157fa9fc08393215c26aa1dc8ad0c67facdf4c7ca8df2acd086