Fix logger using > instead of <
Before this we were comparing and selecting the log with the most sync'd
bytes, vs the least sync'd bytes as is described in the code elsewhere. This
changes it to use the correct comparison and select the one with the least.

FossilOrigin-Name: f0188d377b5af0078b64696da129eb6747509882c89567759ea10f1e6016826e